Aleksei Krutskikh: 2026-06-14: Refactor vector test backend
